requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.1
Django==1.11.5
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.4
gunicorn==19.6.0
idna==2.6
oauthlib==2.0.4
olefile==0.44
Pillow==4.3.0
psycopg2==2.6.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
whitenoise==3.2

social-auth-core==1.5.0
urllib3==1.22
whitenoise==3.2